Draw plausible values (PVs) from an mml fit

Draw plausible values (PVs) from an mml fit

Usage

drawPVs(
  x,
  npv,
  pvVariableNameSuffix = "_dire",
  stochasticBeta = deprecated(),
  ...
)

# S3 method for class 'summary.mmlMeans'
drawPVs(
  x,
  npv = 5L,
  pvVariableNameSuffix = "_dire",
  stochasticBeta = deprecated(),
  ...
)

# S3 method for class 'mmlMeans'
drawPVs(
  x,
  npv = 5L,
  pvVariableNameSuffix = "_dire",
  stochasticBeta = deprecated(),
  ...
)

# S3 method for class 'summary.mmlCompositeMeans'
drawPVs(
  x,
  npv = 20L,
  pvVariableNameSuffix = "_dire",
  stochasticBeta = deprecated(),
  ...
)

# S3 method for class 'mmlCompositeMeans'
drawPVs(
  x,
  npv = 20L,
  pvVariableNameSuffix = "_dire",
  stochasticBeta = deprecated(),
  ...
)

Arguments

x

a fit from a call to mml

npv

integer indicating the number of plausible values to draw

pvVariableNameSuffix

suffix added to new PV variables after construct name and before the plausible value ID. For example, if there is a construct math and the suffix is the default _dire, then the fourth plausible value would have a column name, math_dire4.

stochasticBeta

deprecated.

...

additional parameters passed on to methods.

Value

the stuDat attribute of the mml fit object, x, with columns for the new plausible values.

Details

When the argument passed to stocasticBeta is a data frame then each column is an element that will be used as a regression coefficient for that index of the coefficients vector. The row index used for the nth PV will be the nth row.

Author

Paul Bailey, Sun-joo Lee, and Eric Buehler

Examples

# See Examples in mml